+++ src/siphash24.c
@@ -52,7 +52,7 @@
 ATTRIBUTE((fastcall))
 #endif
 ATTRIBUTE((hot))
-PMOD_EXPORT UINT64 low_hashmem_siphash24( const void *s, size_t len, size_t nbytes,
+PMOD_EXPORT size_t low_hashmem_siphash24( const void *s, size_t len, size_t nbytes,
                                                   UINT64 key )
 {
   const unsigned char * in = (const unsigned char*)s;
